TIVI-2052: Create tizen dbspace directories
There are several middleware packages all racing create the dbspace
directories, and a few that assume that one of the other packages
already created the directories.  This is expressing itself in an
image creation race where the web runtime will sometimes be installed
with a set of critical database files not created.

Since the dbspace directory structure (with /opt/dbspace considered
read-only to all but the preinstalled apps, and /opt/usr/dbspace used
for read-write access) is a fundamental part of the existing tizen
middleware design, then the base filesystem package should be the
thing creating and owning the directores.

Once the new multiuser design has landed then these directories will
likely no longer exist in a shared area.

Change-Id: I0ad7e7f1b8a79a676ddaeb927a844afed9885d9d
Signed-off-by: Rusty Lynch <rusty.lynch@intel.com>