Set the Smack labels on /etc and /var/log so they are shared properly.
The /etc directory contains files (e.g. resolv.conf) that are
updated by System processes but that are read by User processes.
Setting /etc to transmutable System::Shared allows this.

The /var/log directory is used for log files from System and
User services. It gets labeled transmutable System::Log. Note
that the access by System and User are not transmutable. It is
possible that he transmuting will be required by some other
domain in the future.

Change-Id: Iefb0ccce462764415258166e167e1de1d3c6460a
Signed-off-by: Casey Schaufler <casey.schaufler@intel.com>